requirements.txt
voitta
fastapi
uvicorn
python-dotenv
